requirements.txt
boto3==1.26.72
botocore==1.23.24
GitPython==3.1.31
ansible-core==2.12.10
